% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lagger}
\alias{lagger}
\title{Lag a time-series cross-sectional variable}
\usage{
lagger(df, id_var, time_var, weight_name)
}
\arguments{
\item{df}{a data frame containing the unit ID variable and time variables as
well as location' and dependent variables.}

\item{id_var}{a character string identifying the unit ID variable in
\code{df}.}

\item{time_var}{a character string identifying the time variable in
\code{df}. If not specified, then all observations are assumed to be from the
same time period.}

\item{weight_name}{character string providing a custom weighting variable
name.}
}
\description{
Lag a time-series cross-sectional variable
}
