% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monadic_spatial_weights.R
\name{monadic_spatial_weights}
\alias{monadic_spatial_weights}
\title{Find monadic spatial weights for continuous numeric data in a time series data set}
\source{
Neumayer, Eric, and Thomas Plumper. "Making spatial analysis operational:
commands for generating spatial effect variables in monadic and dyadic data."
Stata Journal 10.4 (2010): 585-605.
}
\usage{
monadic_spatial_weights(df, id_var, time_var, location_var, y_var)
}
\arguments{
\item{df}{a data frame containing the unit ID variable and time variables as well as
'location' and dependent variables.}

\item{id_var}{a character string identifying the unit ID variable in \code{df}.}

\item{time_var}{a character string identifying the time variable in \code{df}.}

\item{location_var}{a character string identifying the location of the units in
\code{df}. This is used to create the weighting matrix. Note that the function
finds the relative distance between the units by subtracting their 'location'.}

\item{y_var}{a character string identifying the dependent variable in \code{df}. Note that
an independent variable could also be supplied.}
}
\description{
Find monadic spatial weights for continuous numeric data in a time series data set
}
\details{
Finds spatial effects in monadic data. See Neumayer and Plumper (2010, 591)
for details.
}

